Improve wasm test coverage
