+++ ci/wasm.sh
@@ -9,19 +9,22 @@ GOOS=js GOARCH=wasm go vet ./...
 go install golang.org/x/lint/golint
 GOOS=js GOARCH=wasm golint -set_exit_status ./...
 
-wsEchoOut="$(mktemp -d)/stdout"
-mkfifo "$wsEchoOut"
-go install ./internal/wsecho/cmd/wsecho
-wsecho > "$wsEchoOut" &
+wsjstestOut="$(mktemp -d)/stdout"
+mkfifo "$wsjstestOut"
+go install ./internal/wsjstest
+timeout 30s wsjstest > "$wsjstestOut" &
+wsjstestPID=$!
 
-WS_ECHO_SERVER_URL="$(timeout 10s head -n 1 "$wsEchoOut")" || true
+WS_ECHO_SERVER_URL="$(timeout 10s head -n 1 "$wsjstestOut")" || true
 if [[ -z $WS_ECHO_SERVER_URL ]]; then
-  echo "./internal/wsecho/cmd/wsecho failed to start in 10s"
+  echo "./internal/wsjstest failed to start in 10s"
   exit 1
 fi
 
 go install github.com/agnivade/wasmbrowsertest
 GOOS=js GOARCH=wasm go test -exec=wasmbrowsertest ./... -args "$WS_ECHO_SERVER_URL"
 
-kill %1
-wait -n || true
+if ! wait "$wsjstestPID"; then
+  echo "wsjstest exited unsuccessfully"
+  exit 1
+fi
