+++ internal/wsecho/wsecho.go
@@ -5,33 +5,15 @@ package wsecho
 import (
 	"context"
 	"io"
-	"log"
-	"net/http"
 	"time"
 
 	"nhooyr.io/websocket"
 )
 
-// Serve provides a streaming WebSocket echo server
-// for use in tests.
-func Serve(w http.ResponseWriter, r *http.Request) {
-	c, err := websocket.Accept(w, r, &websocket.AcceptOptions{
-		Subprotocols:       []string{"echo"},
-		InsecureSkipVerify: true,
-	})
-	if err != nil {
-		log.Printf("echo server: failed to accept: %+v", err)
-		return
-	}
-	defer c.Close(websocket.StatusInternalError, "")
-
-	Loop(r.Context(), c)
-}
-
 // Loop echos every msg received from c until an error
 // occurs or the context expires.
 // The read limit is set to 1 << 30.
-func Loop(ctx context.Context, c *websocket.Conn) {
+func Loop(ctx context.Context, c *websocket.Conn) error {
 	defer c.Close(websocket.StatusInternalError, "")
 
 	c.SetReadLimit(1 << 30)
@@ -67,7 +49,7 @@ func Loop(ctx context.Context, c *websocket.Conn) {
 	for {
 		err := echo()
 		if err != nil {
-			return
+			return err
 		}
 	}
 }
